requirements.txt
alembic==0.9.10
amqp==2.3.2
billiard==3.5.0.3
blinker==1.4
celery==4.2.0
click==6.7
configobj==5.0.6
coverage==4.5.1
cymysql==0.9.11
eventlet==0.23.0
Flask==1.0.2
Flask-Celery==2.4.3
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.2.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
greenlet==0.4.14
itsdangerous==0.24
Jinja2==2.10
kombu==4.2.1
Mako==1.0.7
MarkupSafe==1.0
Pillow==5.2.0
python-dateutil==2.7.3
python-editor==1.0.3
pytz==2018.5
redis==2.10.6
six==1.11.0
SQLAlchemy==1.2.9
vine==1.1.4
Werkzeug==0.14.1
WTForms==2.2.1
